Create build-relayer scripts

+++ scripts/ibctest-setup.sh
@@ -3,8 +3,24 @@
 WORKDIR="build/ibc"
 GENFILE="scripts/ibctestchain-gen.sh"
 
+# Ensure desmos is installed
+if ! [ -x "$(which desmos)" ]; then
+  echo "Error: desmos is not installed. Try running 'make install'" >&2
+  exit 1
+fi
+
+# Ensure relayer is installed
+if ! [ -x "$(which rly)" ]; then
+  echo "Error: relayer is not installed." >&2
+  exit 1
+fi
+
 if [ -d $WORKDIR ]; then
     rm -r $WORKDIR
 fi
 
-bash $GENFILE --y
+# generate ibc0 chain config
+bash $GENFILE 0 1 $WORKDIR --y
+
+# generate ibc1 chain config
+bash $GENFILE 1 2 $WORKDIR --y
